adapted code to memory addressing

+++ include/pipeline.h
@@ -12,13 +12,13 @@
 
 typedef struct {
     int inst_id;
-    uint16_t instruction;
+    short int instruction;
     int opcode;
     int r1, r2;
     int8_t imm;
-    int val1, val2; //in case needed later in ALU
-    int result;
-    int pc;
+    int8_t val1, val2; //in case needed later in ALU
+    int8_t result;
+    int8_t pc;
     int valid;
 } PipelineStage;
